[Association] Introduce AssociableTrait with ability to get associated objects by association type
